Move remanence_monitor_event fixture and remove autouse
